Call hour, daily api and map data to view

+++ Weather/Model/Daily.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 struct Daily: Codable {
-    let dt: Int
+    let dt: Double
     let temp: Temp
     let weather: [Weather]
 }
